expose a Session object, let user 'script' it
