+++ docs/pages/session.script.ts
@@ -1,0 +1,5 @@
+import { Session } from '../dep/frugal/frugal_preact.client.ts';
+
+export function main() {
+    new Session();
+}
