+++ packages/frugal_preact/mod.client.ts
@@ -6,3 +6,4 @@ export { hydrate } from './hydrate.tsx';
 export type { App, GetApp } from './hydrate.tsx';
 export { Head } from './Head.tsx';
 export type { HydrationStrategy } from './types.ts';
+export { Session } from './session/Session.ts';
